fix: rewrite deployment test
