+++ justfile
@@ -56,9 +56,9 @@ devloop: unit fmt-imports
 
   just run-storage-docker test-integration
 
-  # just run &
-  # while ! nc -z 127.0.0.1 3000; do sleep 1; done
-  # just test-deployment
+  just run &
+  while ! nc -z 127.0.0.1 3000; do sleep 1; done
+  just test-deployment
 
   echo "✅ Success! ✅"
